backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
umap-learn==0.5.5
scikit-learn==1.4.0
numpy==1.26.3
openai>=1.10.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1